+++ aiohttp/client_ws.py
@@ -7,6 +7,7 @@ import sys
 from ._ws_impl import CLOSED_MESSAGE, WebSocketError, WSMessage, WSMsgType
 
 PY_35 = sys.version_info >= (3, 5)
+PY_352 = sys.version_info >= (3, 5, 2)
 
 
 class ClientWebSocketResponse:
@@ -106,7 +107,7 @@ class ClientWebSocketResponse:
                     self._response.close()
                     return True
 
-                if msg.tp == WSMsgType.CLOSE:
+                if msg.type == WSMsgType.CLOSE:
                     self._close_code = msg.data
                     self._response.close()
                     return True
@@ -139,16 +140,16 @@ class ClientWebSocketResponse:
                     yield from self.close()
                     return WSMessage(WSMsgType.ERROR, exc, None)
 
-                if msg.tp == WSMsgType.CLOSE:
+                if msg.type == WSMsgType.CLOSE:
                     self._closing = True
                     self._close_code = msg.data
                     if not self._closed and self._autoclose:
                         yield from self.close()
                     return msg
                 elif not self._closed:
-                    if msg.tp == WSMsgType.PING and self._autoping:
+                    if msg.type == WSMsgType.PING and self._autoping:
                         self.pong(msg.data)
-                    elif msg.tp == WSMsgType.PONG and self._autoping:
+                    elif msg.type == WSMsgType.PONG and self._autoping:
                         continue
                     else:
                         return msg
@@ -158,17 +159,18 @@ class ClientWebSocketResponse:
     @asyncio.coroutine
     def receive_str(self):
         msg = yield from self.receive()
-        if msg.tp != WSMsgType.TEXT:
+        if msg.type != WSMsgType.TEXT:
             raise TypeError(
-                "Received message {}:{!r} is not str".format(msg.tp, msg.data))
+                "Received message {}:{!r} is not str".format(msg.type,
+                                                             msg.data))
         return msg.data
 
     @asyncio.coroutine
     def receive_bytes(self):
         msg = yield from self.receive()
-        if msg.tp != WSMsgType.BINARY:
+        if msg.type != WSMsgType.BINARY:
             raise TypeError(
-                "Received message {}:{!r} is not bytes".format(msg.tp,
+                "Received message {}:{!r} is not bytes".format(msg.type,
                                                                msg.data))
         return msg.data
 
@@ -178,13 +180,17 @@ class ClientWebSocketResponse:
         return loads(data)
 
     if PY_35:
-        @asyncio.coroutine
-        def __aiter__(self):
-            return self
+        if PY_352:
+            def __aiter__(self):
+                return self
+        else:
+            @asyncio.coroutine
+            def __aiter__(self):
+                return self
 
         @asyncio.coroutine
         def __anext__(self):
             msg = yield from self.receive()
-            if msg.tp == WSMsgType.CLOSE:
+            if msg.type == WSMsgType.CLOSE:
                 raise StopAsyncIteration  # NOQA
             return msg