+++ aiohttp/file_sender.py
@@ -38,12 +38,12 @@ class FileSender:
             fut.set_result(None)
 
     @asyncio.coroutine
-    def _sendfile_system(self, req, resp, fobj, count):
+    def _sendfile_system(self, request, resp, fobj, count):
         """
         Write `count` bytes of `fobj` to `resp` using
         the ``sendfile`` system call.
 
-        `req` should be a :obj:`aiohttp.web.Request` instance.
+        `request` should be a :obj:`aiohttp.web.Request` instance.
 
         `resp` should be a :obj:`aiohttp.web.StreamResponse` instance.
 
@@ -51,15 +51,15 @@ class FileSender:
 
         `count` should be an integer > 0.
         """
-        transport = req.transport
+        transport = request.transport
 
         if transport.get_extra_info("sslcontext"):
-            yield from self._sendfile_fallback(req, resp, fobj, count)
+            yield from self._sendfile_fallback(request, resp, fobj, count)
             return
 
         yield from resp.drain()
 
-        loop = req.app.loop
+        loop = request.app.loop
         # See https://github.com/KeepSafe/aiohttp/issues/958 for details
         out_socket = transport.get_extra_info("socket").dup()
         out_fd = out_socket.fileno()
@@ -74,7 +74,7 @@ class FileSender:
             out_socket.close()
 
     @asyncio.coroutine
-    def _sendfile_fallback(self, req, resp, fobj, count):
+    def _sendfile_fallback(self, request, resp, fobj, count):
         """
         Mimic the :meth:`_sendfile_system` method, but without using the
         ``sendfile`` system call. This should be used on systems that don't
@@ -87,15 +87,13 @@ class FileSender:
         chunk_size = self._chunk_size
 
         chunk = fobj.read(chunk_size)
-        while chunk and count > chunk_size:
+        while True:
             resp.write(chunk)
             yield from resp.drain()
             count = count - chunk_size
-            chunk = fobj.read(chunk_size)
-
-        if chunk:
-            resp.write(chunk[:count])
-            yield from resp.drain()
+            if count <= 0:
+                break
+            chunk = fobj.read(count)
 
     if hasattr(os, "sendfile"):  # pragma: no cover
         _sendfile = _sendfile_system
@@ -103,10 +101,10 @@ class FileSender:
         _sendfile = _sendfile_fallback
 
     @asyncio.coroutine
-    def send(self, req, filepath):
+    def send(self, request, filepath):
         st = filepath.stat()
 
-        modsince = req.if_modified_since
+        modsince = request.if_modified_since
         if modsince is not None and st.st_mtime <= modsince.timestamp():
             from .web_exceptions import HTTPNotModified
             raise HTTPNotModified()
@@ -126,10 +124,10 @@ class FileSender:
         resp.content_length = file_size
         resp.set_tcp_cork(True)
         try:
-            yield from resp.prepare(req)
+            yield from resp.prepare(request)
 
             with filepath.open('rb') as f:
-                yield from self._sendfile(req, resp, f, file_size)
+                yield from self._sendfile(request, resp, f, file_size)
 
         finally:
             resp.set_tcp_nodelay(True)