+++ tests/test_web_websocket_functional.py
@@ -133,11 +133,11 @@ def test_send_recv_text(create_app_and_client, loop):
     ws = yield from client.ws_connect('/')
     ws.send_str('ask')
     msg = yield from ws.receive()
-    assert msg.tp == aiohttp.WSMsgType.TEXT
+    assert msg.type == aiohttp.WSMsgType.TEXT
     assert 'ask/answer' == msg.data
 
     msg = yield from ws.receive()
-    assert msg.tp == aiohttp.WSMsgType.CLOSE
+    assert msg.type == aiohttp.WSMsgType.CLOSE
     assert msg.data == 1000
     assert msg.extra == ''
 
@@ -169,11 +169,11 @@ def test_send_recv_bytes(create_app_and_client, loop):
     ws = yield from client.ws_connect('/')
     ws.send_bytes(b'ask')
     msg = yield from ws.receive()
-    assert msg.tp == aiohttp.WSMsgType.BINARY
+    assert msg.type == aiohttp.WSMsgType.BINARY
     assert b'ask/answer' == msg.data
 
     msg = yield from ws.receive()
-    assert msg.tp == aiohttp.WSMsgType.CLOSE
+    assert msg.type == aiohttp.WSMsgType.CLOSE
     assert msg.data == 1000
     assert msg.extra == ''
 
@@ -205,14 +205,57 @@ def test_send_recv_json(create_app_and_client, loop):
     ws.send_str('{"request": "test"}')
     msg = yield from ws.receive()
     data = msg.json()
-    assert msg.tp == aiohttp.WSMsgType.TEXT
+    assert msg.type == aiohttp.WSMsgType.TEXT
     assert data['response'] == 'test'
 
     msg = yield from ws.receive()
-    assert msg.tp == aiohttp.WSMsgType.CLOSE
+    assert msg.type == aiohttp.WSMsgType.CLOSE
     assert msg.data == 1000
     assert msg.extra == ''
 
     yield from ws.close()
 
     yield from closed
+
+
+@pytest.mark.run_loop
+def test_close_timeout(create_app_and_client, loop):
+    closed = helpers.create_future(loop)
+
+    @asyncio.coroutine
+    def handler(request):
+        ws = web.WebSocketResponse(timeout=0.1)
+        yield from ws.prepare(request)
+        assert 'request' == (yield from ws.receive_str())
+        ws.send_str('reply')
+        begin = ws._loop.time()
+        yield from ws.close()
+        elapsed = ws._loop.time() - begin
+        assert elapsed < 0.201, \
+            'close() should have returned before ' \
+            'at most 2x timeout.'
+        closed.set_result(1)
+        return ws
+
+    app, client = yield from create_app_and_client()
+    app.router.add_route('GET', '/', handler)
+
+    ws = yield from client.ws_connect('/')
+    ws.send_str('request')
+    assert 'reply' == (yield from ws.receive_str())
+
+    # The server closes here.  Then the client sends bogus messages with an
+    # internval shorter than server-side close timeout, to make the server
+    # hanging indefinitely.
+    yield from asyncio.sleep(0.04, loop=loop)
+    ws.send_str('hang')
+    yield from asyncio.sleep(0.04, loop=loop)
+    ws.send_str('hang')
+    yield from asyncio.sleep(0.04, loop=loop)
+    ws.send_str('hang')
+    yield from asyncio.sleep(0.04, loop=loop)
+    ws.send_str('hang')
+    yield from asyncio.sleep(0.04, loop=loop)
+    # The server should have been closed now.
+    assert 1 == (yield from closed)
+    yield from ws.close()